+++ lab 1/lab 1/Tree.h
@@ -82,7 +82,7 @@ int Tree<T>::find_ostov(Graph<T> &graph)
 template<typename T>
 Tree<T>::~Tree()
 {
-	for (int i = 0; i < vertex.size(); ++i)
+	for (size_t i = 0; i < vertex.size(); ++i)
 		delete vertex[i];
 	vertex.clear();
 }
